datashare: splitting the validation funcs
